% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial.R
\name{sig_buffer}
\alias{sig_buffer}
\title{Buffer GIS tools}
\usage{
sig_buffer(capa_entrada, distancia, dissolve = TRUE)
}
\arguments{
\item{capa_entrada}{un objeto de tipo vectorial de la clase sf}

\item{distancia}{Distancia de vecindad para determinar la zona de influencia en m}

\item{dissolve}{Parámetro para controlar la fusión de geometrias.}

\item{campo}{Parámetro para fusionar por un atributo. Opcional.}
}
\value{
Un objeto vectorial de la case sfc_POLYGON
}
\description{
\code{sig_buffer} devuelve la zona de influencia entorno a las geometrias de un shapefile
}
\examples{
\dontrun
sig_buffer(capa_entrada,1000)
sig_buffer(capa_entrada,1000, dissolve = TRUE)
sig_buffer(capa_entrada,1000, dissolve = TRUE, 'Especie')
}
